fix(angle solver): fix net speed

+++ src/DataLink/SerialPort.cpp
@@ -119,17 +119,21 @@ class SerialPort final : public HubHelper<caf::event_based_actor, SerialPortSett
             const HeadInfo infoUp{ SynchronizedClock::instance().now(),
                                    decltype(HeadInfo::transform){ glm::lookAtRH(
                                        glm::dvec3{ 0.0, mConfig.headHeightOffset1, 0.0 },
-                                       glm::dvec3{ std::cos(fdb.yaw + glm::half_pi<double>()) * std::cos(fdb.pitch),
-                                                   mConfig.headHeightOffset1 + std::sin(fdb.pitch),
-                                                   -std::sin(fdb.yaw + glm::half_pi<double>()) * std::cos(fdb.pitch) },
+                                       glm::dvec3{ std::cos(static_cast<double>(fdb.yaw) + glm::half_pi<double>()) *
+                                                       std::cos(static_cast<double>(fdb.pitch)),
+                                                   mConfig.headHeightOffset1 + std::sin(static_cast<double>(fdb.pitch)),
+                                                   -std::sin(static_cast<double>(fdb.yaw) + glm::half_pi<double>()) *
+                                                       std::cos(static_cast<double>(fdb.pitch)) },
                                        glm::dvec3{ 0.0, 1.0, 0.0 }) },
                                    0.0, 0.0 };
             const HeadInfo infoDown{ SynchronizedClock::instance().now(),
                                      decltype(HeadInfo::transform){ glm::lookAtRH(
                                          glm::dvec3{ 0.0, mConfig.headHeightOffset2, 0.0 },
-                                         glm::dvec3{ std::cos(fdb.downYaw + glm::half_pi<double>()) * std::cos(fdb.downPitch),
-                                                     mConfig.headHeightOffset2 + std::sin(fdb.downPitch),
-                                                     -std::sin(fdb.downYaw + glm::half_pi<double>()) * std::cos(fdb.downPitch) },
+                                         glm::dvec3{ std::cos(static_cast<double>(fdb.downYaw) + glm::half_pi<double>()) *
+                                                         std::cos(static_cast<double>(fdb.downPitch)),
+                                                     mConfig.headHeightOffset2 + std::sin(static_cast<double>(fdb.downPitch)),
+                                                     -std::sin(static_cast<double>(fdb.downYaw) + glm::half_pi<double>()) *
+                                                         std::cos(static_cast<double>(fdb.downPitch)) },
                                          glm::dvec3{ 0.0, 1.0, 0.0 }) },
                                      0.0, 0.0 };
